Add files via upload
This is a complete guide on how to boot Linux(OS) on the UETRV_PCORE. This has been authored by Abdullah Azhar, Wajid Ali, Masooma Zia, and Uneeb Kamal. The guide has brought down each necessary step as well as the instructions required to be executed at each step in chronological order.